Abstraction over Persistent and Persisted corfu tables in tests
