+++ runtime/src/main/java/org/corfudb/runtime/collections/DiskBackedCorfuTable.java
@@ -3,6 +3,7 @@ package org.corfudb.runtime.collections;
 import com.google.common.collect.Streams;
 import com.google.common.hash.HashFunction;
 import com.google.common.hash.Hashing;
+import com.google.common.reflect.TypeToken;
 import io.micrometer.core.instrument.Timer;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
@@ -177,6 +178,10 @@ public class DiskBackedCorfuTable<K, V> implements
     public DiskBackedCorfuTable(@NonNull PersistenceOptions persistenceOptions,
                                 @NonNull ISerializer serializer) {
         this(persistenceOptions, defaultOptions, serializer);
+    }
+
+    public static <K, V> TypeToken<DiskBackedCorfuTable<K, V>> getTypeToken() {
+        return new TypeToken<DiskBackedCorfuTable<K, V>>() {};
     }
 
     /**
